Language.Haskell.Tools.AST

Description

A custom AST representation for Haskell tools. Different layers of the AST are recursive, to separate them into modules we introduced source imports.

Documentation

module Language.Haskell.Tools.AST.References

module Language.Haskell.Tools.AST.Helpers

data UModule dom stage

The representation of a haskell module, that is a separate compilation unit. It may or may not have a header.

Constructors

UModule

Fields

_filePragmas :: AnnListG UFilePragma dom stage

_modHead :: AnnMaybeG UModuleHead dom stage

_modImports :: AnnListG UImportDecl dom stage

_modDecl :: AnnListG UDecl dom stage

data UModuleHead dom stage

Module declaration with name and (optional) exports

Constructors

UModuleHead

Fields

_mhName :: Ann UModuleName dom stage

_mhPragma :: AnnMaybeG UModulePragma dom stage

_mhExports :: AnnMaybeG UExportSpecs dom stage

data UExportSpecs dom stage

A list of export specifications surrounded by parentheses

Constructors

UExportSpecs

Fields

_espExports :: AnnListG UExportSpec dom stage

data UExportSpec dom stage

Export specifier

Constructors

UDeclExport
Export a name and related names

Fields

_exportDecl :: Ann UIESpec dom stage

UModuleExport
The export of an imported module ( module A )

Fields

_exportModuleName :: Ann UModuleName dom stage

data UIESpec dom stage

Marks a name to be imported or exported with related names (subspecifier)

Constructors

UIESpec

Fields

_ieModifier :: AnnMaybeG UImportModifier dom stage

_ieName :: Ann UName dom stage

_ieSubspec :: AnnMaybeG USubSpec dom stage

data UImportModifier dom stage

Specifies the imported element

Constructors

UImportPattern
pattern: modifier for importing pattern synonyms

UImportType
type: modifier for importing types

data USubSpec dom stage

Marks how related names will be imported or exported with a given name

Constructors

USubSpecAll
(..): a class exported with all of its methods, or a datatype exported with all of its constructors.

USubSpecList
(a,b,c): a class exported with some of its methods, or a datatype exported with some of its constructors.

Fields

_essList :: AnnListG UName dom stage

data UFilePragma dom stage

Pragmas that must be used before defining the module

Constructors

ULanguagePragma
LANGUAGE pragma, listing the enabled language extensions in that file

Fields

_lpPragmas :: AnnListG ULanguageExtension dom stage

UOptionsPragma
OPTIONS pragma, possibly qualified with a tool, e.g. OPTIONS_GHC

Fields

_opStr :: Ann UStringNode dom stage

data UModulePragma dom stage

Pragmas that must be used after the module head

Constructors

UModuleWarningPragma
A warning pragma attached to the module

Fields

_modWarningStr :: AnnListG UStringNode dom stage

UModuleDeprecatedPragma
A deprecated pragma attached to the module

Fields

_modDeprecatedPragma :: AnnListG UStringNode dom stage

data ULanguageExtension dom stage

The name of the enabled language extension, for example ( LambdaCase )

Constructors

ULanguageExtension

Fields

_langExt :: String

data UImportDecl dom stage

An import declaration: import Module.Name

Constructors

UImportDecl

Fields

_importSource :: AnnMaybeG UImportSource dom stage

_importQualified :: AnnMaybeG UImportQualified dom stage

_importSafe :: AnnMaybeG UImportSafe dom stage

_importPkg :: AnnMaybeG UStringNode dom stage

_importModule :: Ann UModuleName dom stage

_importAs :: AnnMaybeG UImportRenaming dom stage

_importSpec :: AnnMaybeG UImportSpec dom stage

data UImportSpec dom stage

Restriction on the imported names

Constructors

UImportSpecList
Restrict the import definition to ONLY import the listed names

Fields

_importSpecList :: AnnListG UIESpec dom stage

UImportSpecHiding
Restrict the import definition to DONT import the listed names

Fields

_importSpecHiding :: AnnListG UIESpec dom stage

data UImportQualified dom stage

Marks the import as qualified: qualified

Constructors

UImportQualified

data UImportSource dom stage

Marks the import as source: {--}

Constructors

UImportSource

data UImportSafe dom stage

Marks the import as safe: safe

Constructors

UImportSafe

data UTypeNamespace dom stage

Marks an imported name to belong to the type namespace: type

Constructors

UTypeNamespace

data UImportRenaming dom stage

Renaming imports ( as A )

Constructors

UImportRenaming

Fields

_importRename :: Ann UModuleName dom stage

data UModuleName dom stage

The name of a module

Constructors

UModuleName

Fields

_moduleNameString :: String

data TypeKeyword dom stage

The type keyword used to qualify that the type and not the constructor of the same name is referred

Constructors

TypeKeyword

data USplice dom stage

A template haskell splice

Constructors

UIdSplice
A simple name splice: $generateX

Fields

_spliceId :: Ann UName dom stage

UParenSplice
A splice with parentheses: $(generate input)

Fields

_spliceExpr :: Ann UExpr dom stage

data UQuasiQuote dom stage

Template haskell quasi-quotation: [quoter|str]

Constructors

UQuasiQuote

Fields

_qqExprName :: Ann UName dom stage

_qqExprBody :: Ann QQString dom stage

data QQString dom stage

Template Haskell Quasi-quotation content

Constructors

QQString

Fields

_qqString :: String

data UBracket dom stage

Template Haskell bracket expressions

Constructors

UExprBracket
Expression bracket ( [| x + y |] )

Fields

_bracketExpr :: Ann UExpr dom stage

UPatternBracket
Pattern bracket ( [p| Point x y |] )

Fields

_bracketPattern :: Ann UPattern dom stage

UTypeBracket
Type bracket ( [t| (Int,Int) |] )

Fields

_bracketType :: Ann UType dom stage

UDeclsBracket
Declaration bracket ( [d| f :: Int -> Int; f x = x*x |] )

Fields

_bracketDecl :: AnnListG UDecl dom stage

Declarations

data UDecl dom stage

Haskell declarationw

Constructors

UTypeDecl
A type synonym ( type String = [Char] )

Fields

_declHead :: Ann UDeclHead dom stage

_declType :: Ann UType dom stage

UTypeFamilyDecl
A type family declaration ( type family F x )

Fields

_declTypeFamily :: Ann UTypeFamily dom stage

UClosedTypeFamilyDecl
A closed type family declaration

Fields

_declHead :: Ann UDeclHead dom stage

_declSpec :: AnnMaybeG UTypeFamilySpec dom stage

_declDecl :: AnnListG UTypeEqn dom stage

cannot be empty

UDataDecl
A data or newtype declaration. Empty data type declarations without where keyword are always belong to DataDecl.

Fields

_declNewtype :: Ann UDataOrNewtypeKeyword dom stage

_declCtx :: AnnMaybeG UContext dom stage

_declHead :: Ann UDeclHead dom stage

_declCons :: AnnListG UConDecl dom stage

_declDeriving :: AnnListG UDeriving dom stage

UGDataDecl
A GADT-style data or newtype declaration.

Fields

_declNewtype :: Ann UDataOrNewtypeKeyword dom stage

_declCtx :: AnnMaybeG UContext dom stage

_declHead :: Ann UDeclHead dom stage

_declKind :: AnnMaybeG UKindConstraint dom stage

_declGadt :: AnnListG UGadtConDecl dom stage

_declDeriving :: AnnListG UDeriving dom stage

UTypeInstDecl
Type family instance declaration ( type instance Fam T = AssignedT )

Fields

_declInstance :: Ann UInstanceRule dom stage

_declAssignedType :: Ann UType dom stage

UDataInstDecl
Data instance declaration ( data instance Fam T = Con1 | Con2 )

Fields

_declNewtype :: Ann UDataOrNewtypeKeyword dom stage

_declInstance :: Ann UInstanceRule dom stage

_declCons :: AnnListG UConDecl dom stage

_declDeriving :: AnnListG UDeriving dom stage

UGDataInstDecl
GADT-style data instance declaration ( data instance Fam T where ... )

Fields

_declNewtype :: Ann UDataOrNewtypeKeyword dom stage

_declInstance :: Ann UInstanceRule dom stage

_declKind :: AnnMaybeG UKindConstraint dom stage

_declGadt :: AnnListG UGadtConDecl dom stage

UClassDecl
Type class declaration ( class X a [where f = ...] )

Fields

_declCtx :: AnnMaybeG UContext dom stage

_declHead :: Ann UDeclHead dom stage

_declFunDeps :: AnnMaybeG UFunDeps dom stage

_declBody :: AnnMaybeG UClassBody dom stage

UInstDecl
Instance declaration ( instance X T [where f = ...] )

Fields

_declOverlap :: AnnMaybeG UOverlapPragma dom stage

_declInstRule :: Ann UInstanceRule dom stage

_declInstDecl :: AnnMaybeG UInstBody dom stage

UPatternSynonymDecl
Pattern synonyms ( pattern Arrow t1 t2 = App "->" [t1, t2] )

Fields

_declPatSyn :: Ann UPatternSynonym dom stage

UDerivDecl
Standalone deriving declaration ( deriving instance X T )

Fields

_declOverlap :: AnnMaybeG UOverlapPragma dom stage

_declInstRule :: Ann UInstanceRule dom stage

UFixityDecl
Fixity declaration ( infixl 5 +, - )

Fields

_declFixity :: Ann UFixitySignature dom stage

UDefaultDecl
Default types ( default (T1, T2) )

Fields

_declTypes :: AnnListG UType dom stage

UTypeSigDecl
Type signature declaration ( f :: Int -> Int )

Fields

_declTypeSig :: Ann UTypeSignature dom stage

UPatTypeSigDecl
Pattern type signature declaration ( pattern Succ :: Int -> Int )

Fields

_declPatTypeSig :: Ann UPatternTypeSignature dom stage

UValueBinding
Function or value binding ( f x = 12 )

Fields

_declValBind :: Ann UValueBind dom stage

UForeignImport
Foreign import ( foreign import _foo :: Int -> IO Int )

Fields

_declCallConv :: Ann UCallConv dom stage

_declSafety :: AnnMaybeG USafety dom stage

_declName :: Ann UName dom stage

_declForeignType :: Ann UType dom stage

UForeignExport
Foreign export ( foreign export ccall _foo :: Int -> IO Int )

Fields

_declCallConv :: Ann UCallConv dom stage

_declName :: Ann UName dom stage

_declForeignType :: Ann UType dom stage

UPragmaDecl
Top-level pragmas

Fields

_declPragma :: Ann UTopLevelPragma dom stage

URoleDecl
Role annotations ( type role Ptr representational )

Fields

_declRoleType :: Ann UQualifiedName dom stage

_declRoles :: AnnListG URole dom stage

USpliceDecl
A Template Haskell splice declaration ( $(generateDecls) )

Fields

_declSplice :: Ann USplice dom stage

data UDeclHead dom stage

Constructors

UDeclHead
Type or class name

Fields

_dhName :: Ann UName dom stage

UDHParen
Parenthesized type

Fields

_dhBody :: Ann UDeclHead dom stage

UDHApp
Type application

Fields

_dhAppFun :: Ann UDeclHead dom stage

_dhAppOperand :: Ann UTyVar dom stage

UDHInfix
Infix application of the type/class name to the left operand

Fields

_dhLeft :: Ann UTyVar dom stage

_dhOperator :: Ann UOperator dom stage

_dhRight :: Ann UTyVar dom stage

Type class definitions

data UClassBody dom stage

The list of declarations that can appear in a typeclass

Constructors

UClassBody

Fields

_cbElements :: AnnListG UClassElement dom stage

data UClassElement dom stage

Members of a class declaration

Constructors

UClsSig
Signature: f :: A -> B

Fields

_ceTypeSig :: Ann UTypeSignature dom stage

UClsFixity
Fixity signature in class: infixl 1 >>-

Fields

_clsFixity :: Ann UFixitySignature dom stage

UClsDef
Default binding: f x = "aaa"

Fields

_ceBind :: Ann UValueBind dom stage

UClsTypeFam
Declaration of an associated type synonym: type T x :: *

Fields

_ceTypeFam :: Ann UTypeFamily dom stage

UClsTypeDef
Default choice for type synonym: type T x = TE or type instance T x = TE

Fields

_ceHead :: Ann UDeclHead dom stage

_ceKind :: Ann UType dom stage

UClsDefSig
Default signature (by using DefaultSignatures): default _enum :: (Generic a, GEnum (Rep a)) => [a]

Fields

_ceName :: Ann UName dom stage

_ceType :: Ann UType dom stage

UClsMinimal
Minimal pragma: {--}

Fields

_pragmaFormula :: Ann UMinimalFormula dom stage

UClsInline
Inline-like pragma in class definition not supported yet (GHC 8.0.1) | UClsPatSig { _cePatSig :: Ann UPatternTypeSignature dom stage } -- ^ Pattern signature in a class declaration (by using PatternSynonyms)

Fields

_clsInline :: Ann UInlinePragma dom stage

Type class instances

data UInstanceRule dom stage

The instance declaration rule, which is, roughly, the part of the instance declaration before the where keyword.

Constructors

UInstanceRule
Instance head as an instance rule ( X a => Y a )

Fields

_irVars :: AnnMaybeG (AnnListG UTyVar) dom stage

_irCtx :: AnnMaybeG UContext dom stage

_irHead :: Ann UInstanceHead dom stage

data UInstanceHead dom stage

The specification of the class instance declaration

Constructors

UInstanceHeadCon
Type or class name

Fields

_ihConName :: Ann UName dom stage

UInstanceHeadInfix
Infix application of the type/class name to the left operand

Fields

_ihLeftOp :: Ann UType dom stage

_ihOperator :: Ann UOperator dom stage

UInstanceHeadParen
Parenthesized instance head

Fields

_ihHead :: Ann UInstanceHead dom stage

UInstanceHeadApp
Application to one more type

Fields

_ihFun :: Ann UInstanceHead dom stage

_ihType :: Ann UType dom stage

data UInstBody dom stage

Instance body is the implementation of the class functions ( where a x = 1; b x = 2 )

Constructors

UInstBody

Fields

_instBodyDecls :: AnnListG UInstBodyDecl dom stage

data UInstBodyDecl dom stage

Declarations inside an instance declaration.

Constructors

UInstBodyNormalDecl
A normal value binding ( f x = 12 )

Fields

_instBodyDeclFunbind :: Ann UValueBind dom stage

UInstBodyTypeSig
Type signature in instance definition with InstanceSigs

Fields

_instBodyTypeSig :: Ann UTypeSignature dom stage

UInstBodyTypeDecl
An associated type definition ( type A X = B )

Fields

_instBodyTypeEqn :: Ann UTypeEqn dom stage

UInstBodyDataDecl
An associated data type implementation ( data A X = C1 | C2 )

Fields

_instBodyDataNew :: Ann UDataOrNewtypeKeyword dom stage

_instBodyLhsType :: Ann UInstanceRule dom stage

_instBodyDataCons :: AnnListG UConDecl dom stage

_instBodyDerivings :: AnnListG UDeriving dom stage

UInstBodyGadtDataDecl
An associated data type implemented using GADT style

Fields

_instBodyDataNew :: Ann UDataOrNewtypeKeyword dom stage

_instBodyLhsType :: Ann UInstanceRule dom stage

_instBodyDataKind :: AnnMaybeG UKindConstraint dom stage

_instBodyGadtCons :: AnnListG UGadtConDecl dom stage

_instBodyDerivings :: AnnListG UDeriving dom stage

USpecializeInstance
Specialize instance pragma (no phase selection is allowed)

Fields

_specializeInstanceType :: Ann UType dom stage

UInlineInstance
Inline-like pragma in a class instance

Fields

_instanceInline :: Ann UInlinePragma dom stage

UInstanceSpecialize
Specialize pragma not supported yet | UInstBodyPatSyn { _instBodyPatSyn :: Ann UPatternSynonym dom stage } -- ^ A pattern synonym in a class instance

Fields

_specializeInstance :: Ann USpecializePragma dom stage

data UOverlapPragma dom stage

Overlap pragmas. Can be applied to class declarations and class instance declarations.

Constructors

UEnableOverlap
OVERLAP pragma

UDisableOverlap
NO_OVERLAP pragma

UOverlappable
OVERLAPPABLE pragma

UOverlapping
OVERLAPPING pragma

UOverlaps
OVERLAPS pragma

UIncoherentOverlap
INCOHERENT pragma

Type families

data UTypeFamily dom stage

Open type and data families

Constructors

UTypeFamily
Type family declaration ( type family A a :: * -> * )

Fields

_tfHead :: Ann UDeclHead dom stage

_tfSpec :: AnnMaybeG UTypeFamilySpec dom stage

UDataFamily
Data family declaration ( data family A a :: * -> * )

Fields

_tfHead :: Ann UDeclHead dom stage

_tfKind :: AnnMaybeG UKindConstraint dom stage

data UTypeFamilySpec dom stage

Type family specification with kinds specification and injectivity.

Constructors

UTypeFamilyKind
Specifies the kind of a type family ( :: * -> * )

Fields

_tfSpecKind :: Ann UKindConstraint dom stage

UTypeFamilyTyVar
Specifies the kind of a type family ( = t :: * -> * )

Fields

_tfTypeVar :: Ann UTyVar dom stage

UTypeFamilyInjectivity
Specifies the injectivity of a type family ( = r | r -> a )

Fields

_tfInjectivity :: Ann UInjectivityAnn dom stage

data UInjectivityAnn dom stage

Injectivity annotation for type families ( = r | r -> a )

Constructors

UInjectivityAnn

Fields

_injAnnRes :: Ann UTyVar dom stage

_injAnnDeps :: AnnListG UName dom stage

data UTypeEqn dom stage

Type equations as found in closed type families ( T A = S )

Constructors

UTypeEqn

Fields

_teLhs :: Ann UType dom stage

_teRhs :: Ann UType dom stage

Type definitions

data UGadtConDecl dom stage

GADT constructor declaration ( D1 :: { val :: Int } -> T String )

Constructors

UGadtConDecl

Fields

_gadtConNames :: AnnListG UName dom stage

_gadtConTypeArgs :: AnnListG UTyVar dom stage

_gadtConTypeCtx :: AnnMaybeG UContext dom stage

_gadtConType :: Ann UGadtConType dom stage

data UDataOrNewtypeKeyword dom stage

The data or the newtype keyword to define ADTs.

Constructors

UDataKeyword

UNewtypeKeyword

data UGadtConType dom stage

Type of GADT constructors (can be record types: { val :: Int })

Constructors

UGadtNormalType

Fields

_gadtConNormalType :: Ann UType dom stage

UGadtRecordType

Fields

_gadtConRecordFields :: AnnListG UFieldDecl dom stage

_gadtConResultType :: Ann UType dom stage

data UFunDeps dom stage

A list of functional dependencies: | a -> b, c -> d separated by commas

Constructors

UFunDeps

Fields

_funDeps :: AnnListG UFunDep dom stage

data UFunDep dom stage

A functional dependency, given on the form l1 ... ln -> r1 ... rn

Constructors

UFunDep

Fields

_funDepLhs :: AnnListG UName dom stage

_funDepRhs :: AnnListG UName dom stage

data UConDecl dom stage

A constructor declaration for a datatype

Constructors

UConDecl
Ordinary data constructor ( C t1 t2 )

Fields

_conTypeArgs :: AnnListG UTyVar dom stage

_conTypeCtx :: AnnMaybeG UContext dom stage

_conDeclName :: Ann UName dom stage

_conDeclArgs :: AnnListG UType dom stage

URecordDecl
Record data constructor ( C { _n1 :: t1, _n2 :: t2 } )

Fields

_conTypeArgs :: AnnListG UTyVar dom stage

_conTypeCtx :: AnnMaybeG UContext dom stage

_conDeclName :: Ann UName dom stage

_conDeclFields :: AnnListG UFieldDecl dom stage

UInfixConDecl
Infix data constructor ( t1 :+: t2 )

Fields

_conTypeArgs :: AnnListG UTyVar dom stage

_conTypeCtx :: AnnMaybeG UContext dom stage

_conDeclLhs :: Ann UType dom stage

_conDeclOp :: Ann UOperator dom stage

_conDeclRhs :: Ann UType dom stage

data UFieldDecl dom stage

Field declaration ( fld :: Int )

Constructors

UFieldDecl

Fields

_fieldNames :: AnnListG UName dom stage

_fieldType :: Ann UType dom stage

data UDeriving dom stage

A deriving clause following a data type declaration. ( deriving Show or deriving (Show, Eq) )

Constructors

UDerivingOne

Fields

_deriveStrategy :: AnnMaybeG UDeriveStrategy dom stage

_oneDerived :: Ann UInstanceHead dom stage

UDerivings

Fields

_deriveStrategy :: AnnMaybeG UDeriveStrategy dom stage

_allDerived :: AnnListG UInstanceHead dom stage

data UDeriveStrategy dom stage

Constructors

UStockStrategy

UAnyClassStrategy

UNewtypeStrategy

Pattern synonyms

data UPatternTypeSignature dom stage

Pattern type signature declaration ( pattern Succ :: Int -> Int )

Constructors

UPatternTypeSignature

Fields

_patSigName :: AnnListG UName dom stage

_patSigType :: Ann UType dom stage

data UPatternSynonym dom stage

Pattern synonyms: pattern Arrow t1 t2 = App "->" [t1, t2]

Constructors

UPatternSynonym

Fields

_patLhs :: Ann UPatSynLhs dom stage

_patRhs :: Ann UPatSynRhs dom stage

data UPatSynLhs dom stage

Left hand side of a pattern synonym

Constructors

UNormalPatSyn
A left hand side with a constructor name and arguments ( Arrow t1 t2 )

Fields

_patName :: Ann UName dom stage

_patArgs :: AnnListG UName dom stage

UInfixPatSyn
An infix pattern synonym left-hand side ( t1 :+: t2 )

Fields

_patSynLhs :: Ann UName dom stage

_patSynOp :: Ann UOperator dom stage

_patSynRhs :: Ann UName dom stage

URecordPatSyn
A record-style pattern synonym left-hand side ( Arrow { arrowFrom, arrowTo } )

Fields

_patName :: Ann UName dom stage

_patArgs :: AnnListG UName dom stage

data UPatSynRhs dom stage

Right-hand side of pattern synonym

Constructors

UBidirectionalPatSyn
pattern Int = App Int [] or pattern Int <- App Int [] where Int = App Int []

Fields

_patRhsPat :: Ann UPattern dom stage

_patRhsOpposite :: AnnMaybeG UPatSynWhere dom stage

UOneDirectionalPatSyn
 pattern Int <- App Int []

Fields

_patRhsPat :: Ann UPattern dom stage

data UPatSynWhere dom stage

Where clause of pattern synonym (explicit expression direction)

Constructors

UPatSynWhere

Fields

_patOpposite :: AnnListG UMatch dom stage

Foreign imports

data UCallConv dom stage

Call conventions of foreign functions

Constructors

UStdCall

UCCall

UCPlusPlus

UDotNet

UJvm

UJs

UJavaScript

UCApi

data USafety dom stage

Safety annotations for foreign calls

Constructors

USafe

UThreadSafe

UUnsafe

UInterruptible

Role annotations

data URole dom stage

Role annotations for types

Constructors

UNominal

URepresentational

UPhantom

Pragmas

data UTopLevelPragma dom stage

Top level pragmas

Constructors

URulePragma
A pragma that introduces source rewrite rules ( {--} )

Fields

_pragmaRule :: AnnListG URule dom stage

UDeprPragma
A pragma that marks definitions as deprecated ( {-# DEPRECATED f "f will be replaced by g" )

Fields

_pragmaObjects :: AnnListG UName dom stage

_deprMessage :: AnnListG UStringNode dom stage

UWarningPragma
A pragma that marks definitions as deprecated ( {-# WARNING unsafePerformIO "you should know what you are doing" )

Fields

_pragmaObjects :: AnnListG UName dom stage

_warnMessage :: AnnListG UStringNode dom stage

UAnnPragma
A pragma that annotates a definition with an arbitrary value ( {-# ANN f 42 ) TODO: extract pragmas that appear both in top-level and in instances (inline, inlinable, noinline)

Fields

_annotationSubject :: Ann UAnnotationSubject dom stage

_annotateExpr :: Ann UExpr dom stage

UInlinePragmaDecl

Fields

_pragmaInline :: Ann UInlinePragma dom stage

ULinePragma
A pragma for maintaining line numbers in generated sources ( {--} )

Fields

_pragmaLineNum :: Ann LineNumber dom stage

_pragmaFileName :: AnnMaybeG UStringNode dom stage

USpecializeDecl

Fields

_specializePragma :: Ann USpecializePragma dom stage

UCompletePragma
The COMPLETE pragma is used to inform the pattern match checker that a certain set of patterns is complete ( {--} ).

Fields

_pragmaObjects :: AnnListG UName dom stage

_pragmaSignature :: AnnMaybeG UName dom stage

data USpecializePragma dom stage

Constructors

USpecializePragma
A pragma that tells the compiler that a polymorph function should be optimized for a given type ( {--} )

Fields

_pragmaPhase :: AnnMaybeG UPhaseControl dom stage

_specializeDef :: Ann UName dom stage

_specializeType :: AnnListG UType dom stage

data URule dom stage

A rewrite rule ( "map/map" forall f g xs. map f (map g xs) = map (f.g) xs )

Constructors

URule

Fields

_ruleName :: Ann UStringNode dom stage

User name of the rule

_rulePhase :: AnnMaybeG UPhaseControl dom stage

The compilation phases in which the rule can be applied

_ruleBounded :: AnnListG URuleVar dom stage

Variables bound in the rule

_ruleLhs :: Ann UExpr dom stage

The transformed expression

_ruleRhs :: Ann UExpr dom stage

The resulting expression

data URuleVar dom stage

A variable for a rewrite rule. With or without type signature.

Constructors

URuleVar
A simple rule variable

Fields

_ruleVarName :: Ann UName dom stage

USigRuleVar
A rule variable with signature

Fields

_ruleVarName :: Ann UName dom stage

_ruleVarType :: Ann UType dom stage

data UAnnotationSubject dom stage

Annotation allows you to connect an expression to any declaration.

Constructors

UNameAnnotation
The definition with the given name is annotated

Fields

_annotateName :: Ann UName dom stage

UTypeAnnotation
A type with the given name is annotated

Fields

_annotateName :: Ann UName dom stage

UModuleAnnotation
The whole module is annotated

data UMinimalFormula dom stage

Formulas of minimal annotations declaring which functions should be defined.

Constructors

UMinimalName

Fields

_minimalName :: Ann UName dom stage

UMinimalParen

Fields

_minimalInner :: Ann UMinimalFormula dom stage

UMinimalOr
One of the minimal formulas are needed ( min1 | min2 )

Fields

_minimalOrs :: AnnListG UMinimalFormula dom stage

UMinimalAnd
Both of the minimal formulas are needed ( min1 , min2 )

Fields

_minimalAnds :: AnnListG UMinimalFormula dom stage

data LineNumber dom stage

A line number for a line pragma.

Constructors

LineNumber

Fields

_lineNumber :: Int

data UValueBind dom stage

Value binding for top-level and local bindings

Constructors

USimpleBind
Non-function binding ( v = "12" ) TODO: use one name for a function instead of names in each match

Fields

_valBindPat :: Ann UPattern dom stage

_valBindRhs :: Ann URhs dom stage

_valBindLocals :: AnnMaybeG ULocalBinds dom stage

UFunBind
Function binding ( f 0 = 1; f x = x ). All matches must have the same name.

Fields

_funBindMatches :: AnnListG UMatch dom stage

data UMatch dom stage

Clause of function binding

Constructors

UMatch

Fields

_matchLhs :: Ann UMatchLhs dom stage

_matchRhs :: Ann URhs dom stage

_matchBinds :: AnnMaybeG ULocalBinds dom stage

data UMatchLhs dom stage

Something on the left side of the match

Constructors

UNormalLhs
A match lhs with the function name and parameter names ( f a b )

Fields

_matchLhsName :: Ann UName dom stage

_matchLhsArgs :: AnnListG UPattern dom stage

UInfixLhs
An infix match lhs for an operator ( a + b )

Fields

_matchLhsLhs :: Ann UPattern dom stage

_matchLhsOperator :: Ann UOperator dom stage

_matchLhsRhs :: Ann UPattern dom stage

_matchLhsArgs :: AnnListG UPattern dom stage

data ULocalBinds dom stage

Local bindings attached to a declaration ( where x = 42 )

Constructors

ULocalBinds

Fields

_localBinds :: AnnListG ULocalBind dom stage

data ULocalBind dom stage

Bindings that are enabled in local blocks (where or let).

Constructors

ULocalValBind
A local binding for a value

Fields

_localVal :: Ann UValueBind dom stage

ULocalSignature
A local type signature

Fields

_localSig :: Ann UTypeSignature dom stage

ULocalFixity
A local fixity declaration

Fields

_localFixity :: Ann UFixitySignature dom stage

ULocalInline
A local inline pragma

Fields

_localInline :: Ann UInlinePragma dom stage

data UTypeSignature dom stage

A type signature ( f :: Int -> Int )

Constructors

UTypeSignature

Fields

_tsName :: AnnListG UName dom stage

_tsType :: Ann UType dom stage

Fixities

data UFixitySignature dom stage

A fixity signature ( infixl 5 +, - ).

Constructors

UFixitySignature

Fields

_fixityAssoc :: Ann Assoc dom stage

_fixityPrecedence :: AnnMaybeG Precedence dom stage

_fixityOperators :: AnnListG UOperator dom stage

data Assoc dom stage

Associativity of an operator.

Constructors

AssocNone
non-associative operator (declared with infix)

AssocLeft
left-associative operator (declared with infixl)

AssocRight
right-associative operator (declared with infixr)

data Precedence dom stage

Numeric precedence of an operator

Constructors

Precedence

Fields

_precedenceValue :: Int

data URhs dom stage

Right hand side of a value binding (possible with guards): ( = 3 or | x == 1 = 3; | otherwise = 4 )

Constructors

UUnguardedRhs
An unguarded right-hand-side ( = 3 )

Fields

_rhsExpr :: Ann UExpr dom stage

UGuardedRhss
An unguarded right-hand-side ( | x == 1 = 3; | otherwise = 4 )

Fields

_rhsGuards :: AnnListG UGuardedRhs dom stage

data UGuardedRhs dom stage

A guarded right-hand side of a value binding ( | x > 3 = 2 )

Constructors

UGuardedRhs

Fields

_guardStmts :: AnnListG URhsGuard dom stage

Cannot be empty.

_guardExpr :: Ann UExpr dom stage

data URhsGuard dom stage

Guards for value bindings and pattern matches ( Just v x, v 1 )

Constructors

UGuardBind
A bind statement in a pattern guard ( Just v <- x )

Fields

_guardPat :: Ann UPattern dom stage

_guardRhs :: Ann UExpr dom stage

UGuardLet
A let statement in a pattern guard ( let x = 3 )

Fields

_guardBinds :: AnnListG ULocalBind dom stage

UGuardCheck
An expression to check for a pattern guard

Fields

_guardCheck :: Ann UExpr dom stage

data UInlinePragma dom stage

Pragmas that control how the definitions will be inlined

Constructors

UInlinePragma
A pragma that marks a function for inlining to the compiler ( {--} )

Fields

_inlineConlike :: AnnMaybeG UConlikeAnnot dom stage

_inlinePhase :: AnnMaybeG UPhaseControl dom stage

_inlineDef :: Ann UName dom stage

UNoInlinePragma
A pragma that forbids a function from being inlined by the compiler ( {--} )

Fields

_noInlineDef :: Ann UName dom stage

UInlinablePragma
A pragma that marks a function that it may be inlined by the compiler ( {--} )

Fields

_inlinePhase :: AnnMaybeG UPhaseControl dom stage

_inlinableDef :: Ann UName dom stage

data UConlikeAnnot dom stage

A CONLIKE modifier for an INLINE pragma.

Constructors

UConlikeAnnot

data UPhaseControl dom stage

Controls the activation of a rewrite rule ( [1] )

Constructors

UPhaseControl

Fields

_phaseUntil :: AnnMaybeG PhaseInvert dom stage

_phaseNumber :: AnnMaybeG PhaseNumber dom stage

data PhaseNumber dom stage

Phase number for rewrite rules

Constructors

PhaseNumber

Fields

_phaseNum :: Integer

data PhaseInvert dom stage

A tilde that marks the inversion of the phase number

Constructors

PhaseInvert

data UExpr dom stage

Haskell expressions

Constructors

UVar
A variable or a data constructor ( a )

Fields

_exprName :: Ann UName dom stage

ULit
Literal expression ( 42 )

Fields

_exprLit :: Ann ULiteral dom stage

UInfixApp
An infix operator application ( a + b )

Fields

_exprLhs :: Ann UExpr dom stage

_exprOperator :: Ann UOperator dom stage

_exprRhs :: Ann UExpr dom stage

UPrefixApp
Prefix operator application ( -x )

Fields

_exprOperator :: Ann UOperator dom stage

_exprRhs :: Ann UExpr dom stage

UApp
Function application ( f 4 )

Fields

_exprFun :: Ann UExpr dom stage

_exprArg :: Ann UExpr dom stage

ULambda
Lambda expression ( a b -> a + b )

Fields

_exprBindings :: AnnListG UPattern dom stage

at least one

_exprInner :: Ann UExpr dom stage

ULet
Local binding ( let x = 2; y = 3 in e x y )

Fields

_exprFunBind :: AnnListG ULocalBind dom stage

nonempty

_exprInner :: Ann UExpr dom stage

UIf
If expression ( if a then b else c )

Fields

_exprCond :: Ann UExpr dom stage

_exprThen :: Ann UExpr dom stage

_exprElse :: Ann UExpr dom stage

UMultiIf
Multi way if expressions with MultiWayIf extension ( if | guard1 -> expr1; guard2 -> expr2 )

Fields

_exprIfAlts :: AnnListG UGuardedCaseRhs dom stage

UCase
Pattern matching expression ( case expr of pat1 -> expr1; pat2 -> expr2 )

Fields

_exprCase :: Ann UExpr dom stage

_exprAlts :: AnnListG UAlt dom stage

UDo
Do-notation expressions ( do x <- act1; act2 )

Fields

_doKind :: Ann UDoKind dom stage

_exprStmts :: AnnListG UStmt dom stage

UTuple
Tuple expression ( (e1, e2, e3) )

Fields

_tupleElems :: AnnListG UExpr dom stage

UUnboxedTuple
Unboxed tuple expression ( () )

Fields

_tupleElems :: AnnListG UExpr dom stage

UTupleSection
Tuple section, enabled with TupleSections ( (a,,b) ). One of the elements must be missing.

Fields

_tupleSectionElems :: AnnListG UTupSecElem dom stage

UUnboxedTupSec
Unboxed tuple section enabled with TupleSections ( () ). One of the elements must be missing.

Fields

_tupleSectionElems :: AnnListG UTupSecElem dom stage

UList
List expression: [1,2,3]

Fields

_listElems :: AnnListG UExpr dom stage

UParArray
Parallel array expression: [: 1,2,3 :]

Fields

_listElems :: AnnListG UExpr dom stage

UParen
Parenthesized expression: ( a + b )

Fields

_exprInner :: Ann UExpr dom stage

ULeftSection
Left operator section: (1+)

Fields

_exprLhs :: Ann UExpr dom stage

_exprOperator :: Ann UOperator dom stage

URightSection
Right operator section: (+1)

Fields

_exprOperator :: Ann UOperator dom stage

_exprRhs :: Ann UExpr dom stage

URecCon
Record value construction: Point { x = 3, y = -2 }

Fields

_exprRecName :: Ann UName dom stage

_exprRecFields :: AnnListG UFieldUpdate dom stage

URecUpdate
Record value update: p1 { x = 3, y = -2 }

Fields

_exprInner :: Ann UExpr dom stage

_exprRecFields :: AnnListG UFieldUpdate dom stage

UEnum
Enumeration expression ( [1,3..10] )

Fields

_enumFrom :: Ann UExpr dom stage

_enumThen :: AnnMaybeG UExpr dom stage

_enumTo :: AnnMaybeG UExpr dom stage

UParArrayEnum
Parallel array enumeration ( [: 1,3 .. 10 :] )

Fields

_enumFrom :: Ann UExpr dom stage

_enumThen :: AnnMaybeG UExpr dom stage

_enumToFix :: Ann UExpr dom stage

UListComp
List comprehension ( [ (x, y) | x <- xs | y <- ys ] )

Fields

_compExpr :: Ann UExpr dom stage

_compBody :: AnnListG UListCompBody dom stage

Can only have 1 element without ParallelListComp

UParArrayComp
Parallel array comprehensions [: (x, y) | x <- xs , y <- ys :] enabled by ParallelArrays

Fields

_compExpr :: Ann UExpr dom stage

_compBody :: AnnListG UListCompBody dom stage

Can only have 1 element without ParallelListComp

UTypeSig
Explicit type signature ( x :: Int )

Fields

_exprInner :: Ann UExpr dom stage

_exprSig :: Ann UType dom stage

UExplTypeApp
Explicit type application ( show @Integer (read "5") )

Fields

_exprInner :: Ann UExpr dom stage

_exprType :: Ann UType dom stage

UVarQuote
'x for template haskell reifying of expressions

Fields

_quotedName :: Ann UName dom stage

UTypeQuote
''T for template haskell reifying of types

Fields

_quotedName :: Ann UName dom stage

UBracketExpr
Template haskell bracket expression

Fields

_exprBracket :: Ann UBracket dom stage

USplice
Template haskell splice expression, for example: $(gen a) or $x

Fields

_exprSplice :: Ann USplice dom stage

UQuasiQuoteExpr
Template haskell quasi-quotation: [$quoter|str]

Fields

_exprQQ :: Ann UQuasiQuote dom stage

UExprPragma

Fields

_exprPragma :: Ann UExprPragma dom stage

_innerExpr :: Ann UExpr dom stage

UProc
Arrow definition: proc a -> f -< a+1

Fields

_procPattern :: Ann UPattern dom stage

_procExpr :: Ann UCmd dom stage

UArrowApp
Arrow application: f -< a+1

Fields

_exprLhs :: Ann UExpr dom stage

_arrowAppl :: Ann UArrowAppl dom stage

_exprRhs :: Ann UExpr dom stage

ULamCase
Lambda case ( case 0 -> 1; 1 -> 2 )

Fields

_exprAlts :: AnnListG UAlt dom stage

UStaticPtr
Static pointer expression ( static e ). The inner expression must be closed (cannot have variables bound outside)

Fields

_exprInner :: Ann UExpr dom stage

UUnboxedSum
Unboxed sum expression ( () ). XML expressions omitted

Fields

_exprSumPlaceholdersBefore :: AnnListG UUnboxedSumPlaceHolder dom stage

_exprInner :: Ann UExpr dom stage

_exprSumPlaceholdersAfter :: AnnListG UUnboxedSumPlaceHolder dom stage

data UFieldUpdate dom stage

Field update expressions

Constructors

UNormalFieldUpdate
Update of a field ( x = 1 )

Fields

_fieldName :: Ann UName dom stage

_fieldValue :: Ann UExpr dom stage

UFieldPun
Update the field to the value of the same name ( x )

Fields

_fieldUpdateName :: Ann UName dom stage

UFieldWildcard
Update the fields of the bounded names to their values ( .. ). Must be the last initializer. Cannot be used in a record update expression.

Fields

_fieldWildcard :: Ann UFieldWildcard dom stage

data UFieldWildcard dom stage

Marker for a field wildcard. Only needed to attach semantic information in a type-safe way.

Constructors

FldWildcard

data UTupSecElem dom stage

An element of a tuple section that can be an expression or missing (indicating a value from a parameter)

Constructors

Present
An existing element in a tuple section

Fields

_tupSecExpr :: Ann UExpr dom stage

Missing
A missing element in a tuple section

data UAlt' expr dom stage

Clause of case expression ( Just x -> x + 1 )

Constructors

UAlt

Fields

_altPattern :: Ann UPattern dom stage

_altRhs :: Ann (UCaseRhs' expr) dom stage

_altBinds :: AnnMaybeG ULocalBinds dom stage

type UAlt = UAlt' UExpr

type UCmdAlt = UAlt' UCmd

data UCaseRhs' expr dom stage

Right hand side of a match (possible with guards): ( -> 3 or | x == 1 -> 3; | otherwise -> 4 )

Constructors

UUnguardedCaseRhs
Unguarded right-hand side a pattern match ( -> 3 )

Fields

_rhsCaseExpr :: Ann expr dom stage

UGuardedCaseRhss
Guarded right-hand sides of a pattern match ( | x == 1 -> 3; | otherwise -> 4 )

Fields

_rhsCaseGuards :: AnnListG (UGuardedCaseRhs' expr) dom stage

type UCaseRhs = UCaseRhs' UExpr

type UCmdCaseRhs = UCaseRhs' UCmd

data UGuardedCaseRhs' expr dom stage

A guarded right-hand side of pattern matches binding ( | x > 3 -> 2 )

Constructors

UGuardedCaseRhs

Fields

_caseGuardStmts :: AnnListG URhsGuard dom stage

Cannot be empty.

_caseGuardExpr :: Ann expr dom stage

type UGuardedCaseRhs = UGuardedCaseRhs' UExpr

type UCmdGuardedCaseRhs = UGuardedCaseRhs' UCmd

data UExprPragma dom stage

Pragmas that can be applied to expressions

Constructors

UCorePragma
A CORE pragma for adding notes to expressions.

Fields

_pragmaStr :: Ann UStringNode dom stage

USccPragma
An SCC pragma for defining cost centers for profiling

Fields

_pragmaStr :: Ann UStringNode dom stage

UGeneratedPragma
A pragma that describes if an expression was generated from a code fragment by an external tool ( {--} )

Fields

_pragmaSrcRange :: Ann USourceRange dom stage

data USourceRange dom stage

In-AST source ranges (for generated pragmas)

Constructors

USourceRange

Fields

_srFileName :: Ann UStringNode dom stage

_srFromLine :: Ann Number dom stage

_srFromCol :: Ann Number dom stage

_srToLine :: Ann Number dom stage

_srToCol :: Ann Number dom stage

data Number dom stage

Constructors

Number

Fields

_numberInteger :: Integer

Arrows

data UCmd dom stage

Constructors

UArrowAppCmd
An arrow application command ( f -< x + 1 )

Fields

_cmdLhs :: Ann UExpr dom stage

_cmdArrowOp :: Ann UArrowAppl dom stage

_cmdRhs :: Ann UExpr dom stage

UArrowFormCmd
A form command ( (|untilA (increment -< x+y) (within 0.5 -< x)|) )

Fields

_cmdExpr :: Ann UExpr dom stage

_cmdInnerCmds :: AnnListG UCmd dom stage

UAppCmd
A function application command

Fields

_cmdInnerCmd :: Ann UCmd dom stage

_cmdApplied :: Ann UExpr dom stage

UInfixCmd
An infix command application

Fields

_cmdLeftCmd :: Ann UCmd dom stage

_cmdOperator :: Ann UName dom stage

_cmdRightCmd :: Ann UCmd dom stage

ULambdaCmd
A lambda command

Fields

_cmdBindings :: AnnListG UPattern dom stage

at least one

_cmdInner :: Ann UCmd dom stage

UParenCmd
A parenthesized command

Fields

_cmdInner :: Ann UCmd dom stage

UCaseCmd
A pattern match command

Fields

_cmdExpr :: Ann UExpr dom stage

_cmdAlts :: AnnListG UCmdAlt dom stage

UIfCmd
An if command ( if f x y then g -< x+1 else h -< y+2 )

Fields

_cmdExpr :: Ann UExpr dom stage

_cmdThen :: Ann UCmd dom stage

_cmdElse :: Ann UCmd dom stage

ULetCmd
A local binding command ( let z = x+y )

Fields

_cmdBinds :: AnnListG ULocalBind dom stage

nonempty

_cmdInner :: Ann UCmd dom stage

UDoCmd
A do-notation in a command

Fields

_cmdStmts :: AnnListG UCmdStmt dom stage

data UArrowAppl dom stage

Constructors

ULeftAppl
Left arrow application: -<

URightAppl
Right arrow application: >-

ULeftHighApp
Left arrow high application: -<<

URightHighApp
Right arrow high application: >>-

data UUnboxedSumPlaceHolder dom stage

Constructors

UUnboxedSumPlaceHolder

data UStmt' expr dom stage

Normal monadic statements

Constructors

UBindStmt
Binding statement ( x <- action )

Fields

_stmtPattern :: Ann UPattern dom stage

_stmtExpr :: Ann expr dom stage

UExprStmt
Non-binding statement ( action )

Fields

_stmtExpr :: Ann expr dom stage

ULetStmt
Let statement ( let x = 3; y = 4 )

Fields

_stmtBinds :: AnnListG ULocalBind dom stage

URecStmt
A recursive binding statement with ( rec b <- f a c; c <- f b a )

Fields

_cmdStmtBinds :: AnnListG (UStmt' expr) dom stage

type UStmt = UStmt' UExpr

type UCmdStmt = UStmt' UCmd

data UListCompBody dom stage

Body of a list comprehension: ( | x <- [1..10] )

Constructors

UListCompBody

Fields

_compStmts :: AnnListG UCompStmt dom stage

data UCompStmt dom stage

List comprehension statement

Constructors

UCompStmt
Normal monadic statement of a list comprehension

Fields

_compStmt :: Ann UStmt dom stage

UThenStmt
Then statements by TransformListComp ( then sortWith by (x + y) )

Fields

_thenExpr :: Ann UExpr dom stage

_byExpr :: AnnMaybeG UExpr dom stage

UGroupStmt
Grouping statements by TransformListComp ( then group by (x + y) using groupWith ) Note: either byExpr or usingExpr must have a value

Fields

_byExpr :: AnnMaybeG UExpr dom stage

_usingExpr :: AnnMaybeG UExpr dom stage

data UDoKind dom stage

Keywords do or mdo to start a do-block

Constructors

UDoKeyword

UMDoKeyword

data UPattern dom stage

Representation of patterns for pattern bindings

Constructors

UVarPat
Pattern name binding

Fields

_patternName :: Ann UName dom stage

ULitPat
Literal pattern

Fields

_patternLiteral :: Ann ULiteral dom stage

UInfixAppPat
Infix constructor application pattern ( a :+: b )

Fields

_patternLhs :: Ann UPattern dom stage

_patternOperator :: Ann UOperator dom stage

_patternRhs :: Ann UPattern dom stage

UAppPat
Constructor application pattern ( Point x y )

Fields

_patternName :: Ann UName dom stage

_patternArgs :: AnnListG UPattern dom stage

UTuplePat
Tuple pattern ( (x,y) )

Fields

_patternElems :: AnnListG UPattern dom stage

UUnboxTuplePat
Unboxed tuple pattern ( () )

Fields

_patternElems :: AnnListG UPattern dom stage

UListPat
List pattern ( [1,2,a,x] )

Fields

_patternElems :: AnnListG UPattern dom stage

UParArrPat
Parallel array pattern ( [:1,2,a,x:] )

Fields

_patternElems :: AnnListG UPattern dom stage

UParenPat
Parenthesised patterns

Fields

_patternInner :: Ann UPattern dom stage

URecPat
Record pattern ( Point { x = 3, y } )

Fields

_patternName :: Ann UName dom stage

_patternFields :: AnnListG UPatternField dom stage

UAsPat
As-pattern (explicit name binding) ( ls@(hd:_) )

Fields

_patternName :: Ann UName dom stage

_patternInner :: Ann UPattern dom stage

UWildPat
Wildcard pattern: ( _ )

UIrrefutablePat
Irrefutable pattern ( ~(x:_) )

Fields

_patternInner :: Ann UPattern dom stage

UBangPat
Bang pattern ( !x )

Fields

_patternInner :: Ann UPattern dom stage

UTypeSigPat
Pattern with explicit type signature ( x :: Int )

Fields

_patternInner :: Ann UPattern dom stage

_patternType :: Ann UType dom stage

UViewPat
View pattern ( f -> Just 1 ) regular list pattern omitted xml patterns omitted

Fields

_patternExpr :: Ann UExpr dom stage

_patternInner :: Ann UPattern dom stage

USplicePat
Splice patterns: $(generateX inp)

Fields

_patternSplice :: Ann USplice dom stage

UQuasiQuotePat
Quasi-quoted patterns: [| 1 + 2 |]

Fields

_patQQ :: Ann UQuasiQuote dom stage

UNPlusKPat

Fields

_patternName :: Ann UName dom stage

_patternLit :: Ann ULiteral dom stage

UUnboxedSumPat
Unboxed sum pattern ( () ).

Fields

_patternSumPlaceholdersBefore :: AnnListG UUnboxedSumPlaceHolder dom stage

_patternInner :: Ann UPattern dom stage

_patternSumPlaceholdersAfter :: AnnListG UUnboxedSumPlaceHolder dom stage

data UPatternField dom stage

Constructors

UNormalFieldPattern
Named field pattern ( p = Point 3 2 )

Fields

_fieldPatternName :: Ann UName dom stage

_fieldPattern :: Ann UPattern dom stage

UFieldPunPattern
Named field pun ( p )

Fields

_fieldPatternName :: Ann UName dom stage

UFieldWildcardPattern
Wildcard field pattern ( .. )

Fields

_fieldPatternWildcard :: Ann UFieldWildcard dom stage

data UTyVar dom stage

Type variable declaration

Constructors

UTyVarDecl

Fields

_tyVarName :: Ann UName dom stage

_tyVarKind :: AnnMaybeG UKindConstraint dom stage

data UType dom stage

Haskell types

Constructors

UTyForall
Forall types ( forall x y . type )

Fields

_typeBounded :: AnnListG UTyVar dom stage

_typeType :: Ann UType dom stage

UTyCtx
Type with a context ( forall x y . type )

Fields

_typeCtx :: Ann UContext dom stage

_typeType :: Ann UType dom stage

UTyFun
Function types ( a -> b )

Fields

_typeParam :: Ann UType dom stage

_typeResult :: Ann UType dom stage

UTyTuple
Tuple types ( (a,b) )

Fields

_typeElements :: AnnListG UType dom stage

UTyUnbTuple
Unboxed tuple types ( () )

Fields

_typeElements :: AnnListG UType dom stage

UTyList
List type with special syntax ( [a] )

Fields

_typeElement :: Ann UType dom stage

UTyParArray
Parallel array type ( [:a:] )

Fields

_typeElement :: Ann UType dom stage

UTyApp
Type application ( F a )

Fields

_typeCon :: Ann UType dom stage

_typeArg :: Ann UType dom stage

UTyVar
Type variable or constructor ( a )

Fields

_typeName :: Ann UName dom stage

UTyParen
Type surrounded by parentheses ( (T a) )

Fields

_typeInner :: Ann UType dom stage

UTyInfix
Infix type constructor ( (a <: b) )

Fields

_typeLeft :: Ann UType dom stage

_typeOperator :: Ann UOperator dom stage

_typeRight :: Ann UType dom stage

UTyKinded
Type with explicit kind signature ( a :: * )

Fields

_typeInner :: Ann UType dom stage

_typeKind :: Ann UKind dom stage

UTyPromoted

Fields

_tpPromoted :: Ann (UPromoted UType) dom stage

UTySplice
A Template Haskell splice type ( $(genType) ).

Fields

_tsSplice :: Ann USplice dom stage

UTyQuasiQuote
A Template Haskell quasi-quote type ( [quoter| ... ] ).

Fields

_typeQQ :: Ann UQuasiQuote dom stage

UTyBang
Strict type marked with !.

Fields

_typeInner :: Ann UType dom stage

UTyLazy
Lazy type marked with ~. (Should only be used if Strict or StrictData language extension is used)

Fields

_typeInner :: Ann UType dom stage

UTyUnpack
Strict type marked with UNPACK pragma. (Usually contains the bang mark.)

Fields

_typeInner :: Ann UType dom stage

UTyNoUnpack
Strict type marked with NOUNPACK pragma. (Usually contains the bang mark.)

Fields

_typeInner :: Ann UType dom stage

UTyWildcard
A wildcard type ( _ ) with -XPartialTypeSignatures

UTyNamedWildc
A named wildcard type ( _t ) with -XPartialTypeSignatures

Fields

_typeWildcardName :: Ann UName dom stage

UUnbSumType
An unboxed sum type ( () )

Fields

_typeElements :: AnnListG UType dom stage

data UContext dom stage

Constructors

UContext
Assertions with the fat arrow ( C a => ... )

Fields

_contextAssertion :: Ann UAssertion dom stage

data UAssertion dom stage

A single assertion in the context

Constructors

UClassAssert
Class assertion (Cls x)

Fields

_assertClsName :: Ann UName dom stage

_assertTypes :: AnnListG UType dom stage

UInfixAssert
Infix class assertion, also contains type equations ( a ~ X y )

Fields

_assertLhs :: Ann UType dom stage

_assertOp :: Ann UOperator dom stage

_assertRhs :: Ann UType dom stage

UImplicitAssert
Assertion for implicit parameter binding ( ?cmp :: a -> a -> Bool )

Fields

_assertImplVar :: Ann UName dom stage

_assertImplType :: Ann UType dom stage

UTupleAssert
Multiple assertions in one ( (Ord a, Show a) )

Fields

_innerAsserts :: AnnListG UAssertion dom stage

UWildcardAssert
Wildcard assertion ( _ ), enabled by PartialTypeSignatures

data UKindConstraint dom stage

Kind constraint ( :: * -> * )

Constructors

UKindConstraint

Fields

_kindConstr :: Ann UKind dom stage

data UKind dom stage

Haskell kinds

Constructors

UStarKind
*, the kind of types

UUnboxKind
#, the kind of unboxed types

UFunKind
->, the kind of type constructor

Fields

_kindLeft :: Ann UKind dom stage

_kindRight :: Ann UKind dom stage

UParenKind
A parenthesised kind

Fields

_kindParen :: Ann UKind dom stage

UVarKind
Kind variable (using PolyKinds extension)

Fields

_kindVar :: Ann UName dom stage

UAppKind
Kind application ( k1 k2 )

Fields

_kindAppFun :: Ann UKind dom stage

_kindAppArg :: Ann UKind dom stage

UInfixAppKind
Infix kind application ( k1 ~> k2 )

Fields

_kindLhs :: Ann UKind dom stage

_kindAppOp :: Ann UOperator dom stage

_kindRhs :: Ann UKind dom stage

UListKind
A list kind ( [k] )

Fields

_kindElem :: Ann UKind dom stage

UTupleKind
A tuple kind ( (Symbol, *) )

Fields

_kindElems :: AnnListG UKind dom stage

UPromotedKind
A promoted kind ( '(k1,k2,k3) )

Fields

_kindPromoted :: Ann (UPromoted UKind) dom stage

UTypeKind
A type on the kind level with TypeInType

Fields

_kindType :: Ann UType dom stage

data UPromoted t dom stage

Constructors

UPromotedInt
Numeric value promoted to the kind level.

Fields

_promotedIntValue :: Integer

UPromotedString
String value promoted to the kind level.

Fields

_promotedStringValue :: String

UPromotedCon
A data constructor value promoted to the kind level.

Fields

_promotedConName :: Ann UName dom stage

UPromotedList
A list of elements as a kind.

Fields

_promotedElements :: AnnListG t dom stage

UPromotedTuple
A tuple of elements as a kind.

Fields

_promotedElements :: AnnListG t dom stage

UPromotedUnit
Kind of the unit value ().

data ULiteral dom stage

Haskell literals

Constructors

UCharLit
Character literal: c

Fields

_charLitValue :: Char

UStringLit
String literal: "abc"

Fields

_stringLitValue :: String

UIntLit
Integer literal: 12

Fields

_intLitValue :: Integer

UFracLit
Fractional literal: 3.14

Fields

_fracLitValue :: Rational

UPrimIntLit
Primitive integer literal (of type Int#): 32#

Fields

_intLitValue :: Integer

UPrimWordLit
Primitive word literal (of type Word#): 32##

Fields

_intLitValue :: Integer

UPrimFloatLit
Primitive float literal (of type Float#): 3.14#

Fields

_floatLitValue :: Rational

UPrimDoubleLit
Primitive double literal (of type Double#): 3.14##

Fields

_floatLitValue :: Rational

UPrimCharLit
Primitive character literal (of type Char#): c#

Fields

_charLitValue :: Char

UPrimStringLit
Primitive string literal (of type Addr#): "xxx"#

Fields

_stringLitValue :: String

data UOperator dom stage

Constructors

UBacktickOp
A normal name used as an operator with backticks: a mod b

Fields

_operatorName :: Ann UQualifiedName dom stage

UNormalOp
A normal operator used as an operator.

Fields

_operatorName :: Ann UQualifiedName dom stage

data UName dom stage

Constructors

UParenName
Parenthesized name: foldl (+) 0

Fields

_simpleName :: Ann UQualifiedName dom stage

UNormalName
A normal, non-operator name.

Fields

_simpleName :: Ann UQualifiedName dom stage

UImplicitName
Implicit name: ?var

Fields

_simpleName :: Ann UQualifiedName dom stage

data UQualifiedName dom stage

Possible qualified names. Contains also implicit names. Linear implicit parameter: %x. Non-linear implicit parameter: ?x.

Constructors

UQualifiedName

Fields

_qualifiers :: AnnListG UNamePart dom stage

_unqualifiedName :: Ann UNamePart dom stage

nameFromList :: AnnListG UNamePart dom stage -> UQualifiedName dom stage

data UNamePart dom stage

Parts of a qualified name.

Constructors

UNamePart

Fields

_simpleNameStr :: String

data UStringNode dom stage

Program elements formatted as string literals (import packages, pragma texts)

Constructors

UStringNode

Fields

_stringNodeStr :: String

module Language.Haskell.Tools.AST.Ann

module Language.Haskell.Tools.AST.SemaInfoClasses