Language.Haskell.Tools.AST.SemaInfoClasses

Documentation

type HasNoSemanticInfo dom si = SemanticInfo dom si ~ NoSemanticInfo

class HasImplicitFieldsInfo' si where

Minimal complete definition

semanticsImplicitFlds

Methods

semanticsImplicitFlds :: si -> [(Name, Name)]

type HasImplicitFieldsInfo dom = (Domain dom, HasImplicitFieldsInfo' (SemanticInfo dom UFieldWildcard))

class HasImportInfo' si where

Minimal complete definition

semanticsImportedModule, semanticsAvailable, semanticsImported, semanticsTransMods

Methods

semanticsImportedModule :: si -> Module

semanticsAvailable :: si -> [Name]

semanticsImported :: si -> [Name]

semanticsTransMods :: si -> [Module]

type HasImportInfo dom = (Domain dom, HasImportInfo' (SemanticInfo dom UImportDecl))

class HasModuleInfo' si where

Minimal complete definition

semanticsModule, semanticsDynFlags, isBootModule, semanticsImplicitImports, semanticsPrelTransMods

Methods

semanticsModule :: si -> Module

semanticsDynFlags :: si -> DynFlags

isBootModule :: si -> Bool

semanticsImplicitImports :: si -> [Name]

semanticsPrelTransMods :: si -> [Module]

type HasModuleInfo dom = (Domain dom, HasModuleInfo' (SemanticInfo dom UModule))

class HasSourceInfoInSema' si where

Minimal complete definition

semanticsSourceInfo

Methods

semanticsSourceInfo :: si -> Maybe SrcSpan

class HasDefiningInfo' si where

Infos that store if they were used to define a name

Minimal complete definition

semanticsDefining

Methods

semanticsDefining :: si -> Bool

type HasDefiningInfo dom = (Domain dom, HasDefiningInfo' (SemanticInfo dom UQualifiedName))

class HasScopeInfo' si where

Infos that contain the names that are available in theirs scope

Minimal complete definition

semanticsScope

Methods

semanticsScope :: si -> Scope

type HasScopeInfo dom = (Domain dom, HasScopeInfo' (SemanticInfo dom UQualifiedName), HasScopeInfo' (SemanticInfo dom UExpr))

class HasFixityInfo' si where

Infos that may have a fixity information

Minimal complete definition

semanticsFixity

Methods

semanticsFixity :: si -> Maybe Fixity

type HasFixityInfo dom = (Domain dom, HasFixityInfo' (SemanticInfo dom UQualifiedName))

class HasNameInfo' si => HasIdInfo' si where

Infos that may have a typed name that can be extracted

Minimal complete definition

semanticsId

Methods

semanticsId :: si -> Id

type HasIdInfo dom = (Domain dom, HasIdInfo' (SemanticInfo dom UQualifiedName))

class HasLiteralInfo' si where

Info of types

Minimal complete definition

semanticsLiteralType

Methods

semanticsLiteralType :: si -> Type

type HasLiteralInfo dom = (Domain dom, HasLiteralInfo' (SemanticInfo dom ULiteral))

Domains that have semantic information for literals

class HasNameInfo' si where

Infos that may have a name that can be extracted

Minimal complete definition

semanticsName

Methods

semanticsName :: si -> Maybe Name

type HasNameInfo dom = (Domain dom, HasNameInfo' (SemanticInfo dom UQualifiedName))

Domains that have semantic information for names

semanticsLitType :: Ann ULiteral IdDom st -> Type

getInstances :: GhcMonad m => [Module] -> m ([ClsInst], [FamInst])

Gets the class and family instances from a module.

data UsageSpec

Constructors

UsageSpec

Fields

usageQualified :: Bool

usageQualifier :: String

usageAs :: String