Language.Haskell.Tools.AST.SemaInfoTypes

Documentation

data NoSemanticInfo

Semantic info type for any node not carrying additional semantic information

data ScopeInfo

Info for expressions that tells which definitions are in scope

data NameInfo n

Info corresponding to a name

data CNameInfo

Info corresponding to a name that is correctly identified

data ModuleInfo n

Info for the module element

data ImportInfo n

Info corresponding to an import declaration

data ImplicitFieldInfo

Info corresponding to an record-wildcard

type Scope = [[(Name, Maybe [UsageSpec], Maybe Name)]]

data UsageSpec

Constructors

UsageSpec

Fields

usageQualified :: Bool

usageQualifier :: String

usageAs :: String

data LiteralInfo

Constructors

LiteralInfo

Fields

_literalType :: Type

data PreLiteralInfo

Constructors

RealLiteralInfo

Fields

_realLiteralType :: Type

PreLiteralInfo

Fields

_preLiteralLoc :: SrcSpan

exprScopedLocals :: Lens ScopeInfo ScopeInfo Scope Scope

nameScopedLocals :: forall n. Lens (NameInfo n) (NameInfo n) Scope Scope

nameIsDefined :: forall n. Lens (NameInfo n) (NameInfo n) Bool Bool

nameInfo :: forall n n'. Partial (NameInfo n) (NameInfo n') n n'

ambiguousName :: forall n. Partial (NameInfo n) (NameInfo n) RdrName RdrName

nameLocation :: forall n. Partial (NameInfo n) (NameInfo n) SrcSpan SrcSpan

implicitName :: forall n. Partial (NameInfo n) (NameInfo n) String String

cnameScopedLocals :: Lens CNameInfo CNameInfo Scope Scope

cnameIsDefined :: Lens CNameInfo CNameInfo Bool Bool

cnameInfo :: Lens CNameInfo CNameInfo Id Id

cnameFixity :: Lens CNameInfo CNameInfo (Maybe Fixity) (Maybe Fixity)

defModuleName :: forall n. Lens (ModuleInfo n) (ModuleInfo n) Module Module

defDynFlags :: forall n. Lens (ModuleInfo n) (ModuleInfo n) DynFlags DynFlags

defIsBootModule :: forall n. Lens (ModuleInfo n) (ModuleInfo n) Bool Bool

implicitNames :: forall n n'. Lens (ModuleInfo n) (ModuleInfo n') [PName n] [PName n']

importedModule :: forall n. Lens (ImportInfo n) (ImportInfo n) Module Module

availableNames :: forall n. Lens (ImportInfo n) (ImportInfo n) [n] [n]

importedNames :: forall n. Lens (ImportInfo n) (ImportInfo n) [PName n] [PName n]

implicitFieldBindings :: Lens ImplicitFieldInfo ImplicitFieldInfo [(Name, Name)] [(Name, Name)]

prelTransMods :: forall n. Lens (ModuleInfo n) (ModuleInfo n) [Module] [Module]

importTransMods :: forall n. Lens (ImportInfo n) (ImportInfo n) [Module] [Module]

literalType :: Lens LiteralInfo LiteralInfo Type Type

mkNoSemanticInfo :: NoSemanticInfo

mkScopeInfo :: Scope -> ScopeInfo

Creates the information about the definitions in scope

mkNameInfo :: Scope -> Bool -> n -> NameInfo n

Creates semantic information for an unambiguous name

mkAmbiguousNameInfo :: Scope -> Bool -> RdrName -> SrcSpan -> NameInfo n

Creates semantic information for a name that is ambiguous because the lack of type info

mkImplicitNameInfo :: Scope -> Bool -> String -> SrcSpan -> NameInfo n

Creates semantic information for an implicit name

mkCNameInfo :: Scope -> Bool -> Id -> Maybe Fixity -> CNameInfo

Create a typed name semantic information

mkModuleInfo :: Module -> DynFlags -> Bool -> [PName n] -> [Module] -> ModuleInfo n

Creates semantic information for the module element. Strict in the list of implicitely imported, orphan and family instances.

mkImportInfo :: Module -> [n] -> [PName n] -> [Module] -> ImportInfo n

Creates semantic information for an import declaration Strict in the list of the used and imported declarations, orphan and family instances.

mkImplicitFieldInfo :: [(Name, Name)] -> ImplicitFieldInfo

Creates semantic information for a wildcard field binding

data PName n

Constructors

PName

Fields

_pName :: n

_pNameParent :: Maybe n

pName :: forall n. Lens (PName n) (PName n) n n

pNameParent :: forall n. Lens (PName n) (PName n) (Maybe n) (Maybe n)

getInstances :: [Module] -> Ghc ([ClsInst], [FamInst])

Gets the class and family instances from a module.

Orphan instances

Data FamInst

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> FamInst -> c FamInst

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c FamInst

toConstr :: FamInst -> Constr

dataTypeOf :: FamInst -> DataType

dataCast1 :: Typeable (* -> *) t => (forall d. Data d => c (t d)) -> Maybe (c FamInst)

dataCast2 :: Typeable (* -> * -> *) t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c FamInst)

gmapT :: (forall b. Data b => b -> b) -> FamInst -> FamInst

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> FamInst -> r

gmapQr :: (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> FamInst -> r

gmapQ :: (forall d. Data d => d -> u) -> FamInst -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> FamInst -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> FamInst -> m FamInst

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> FamInst -> m FamInst

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> FamInst -> m FamInst

Data FamFlavor

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> FamFlavor -> c FamFlavor

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c FamFlavor

toConstr :: FamFlavor -> Constr

dataTypeOf :: FamFlavor -> DataType

dataCast1 :: Typeable (* -> *) t => (forall d. Data d => c (t d)) -> Maybe (c FamFlavor)

dataCast2 :: Typeable (* -> * -> *) t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c FamFlavor)

gmapT :: (forall b. Data b => b -> b) -> FamFlavor -> FamFlavor

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> FamFlavor -> r

gmapQr :: (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> FamFlavor -> r

gmapQ :: (forall d. Data d => d -> u) -> FamFlavor -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> FamFlavor -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> FamFlavor -> m FamFlavor

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> FamFlavor -> m FamFlavor

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> FamFlavor -> m FamFlavor

Data DynFlags

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> DynFlags -> c DynFlags

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c DynFlags

toConstr :: DynFlags -> Constr

dataTypeOf :: DynFlags -> DataType

dataCast1 :: Typeable (* -> *) t => (forall d. Data d => c (t d)) -> Maybe (c DynFlags)

dataCast2 :: Typeable (* -> * -> *) t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c DynFlags)

gmapT :: (forall b. Data b => b -> b) -> DynFlags -> DynFlags

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> DynFlags -> r

gmapQr :: (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> DynFlags -> r

gmapQ :: (forall d. Data d => d -> u) -> DynFlags -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> DynFlags -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> DynFlags -> m DynFlags

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> DynFlags -> m DynFlags

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> DynFlags -> m DynFlags