Language.Haskell.Tools.AST.Ann

Description

Parts of AST representation for keeping extra data

Annotation type resolution

data RangeStage

A stage in which the nodes are marked with the ranges in the source files which contain the source code of the given AST element.

data NormRangeStage

A stage in which the nodes are still marked with ranges, but these ranges are normalized. Optional and list elements also have ranges in that state.

data RngTemplateStage

A stage in which AST elements are marked with templates. These templates are hierarchical, and contain the places of the children elements of the node.

data SrcTemplateStage

A stage where the annotation controls how the original source code can be retrieved from the AST. A source template is assigned to each node. It has holes where the content of an other node should be printed and ranges for the source code of the node.

data Dom name

With this domain, semantic information can be parameterized. In practice it is only used if the compilation cannot proceed past the type checking phase.

data IdDom

type SemanticInfo (domain :: *) (node :: * -> * -> *) = SemanticInfo' domain (SemaInfoClassify node)

data SemaInfoNameCls

data SemaInfoLitCls

data SemaInfoExprCls

data SemaInfoImportCls

data SemaInfoModuleCls

data SemaInfoDefaultCls

data SemaInfoWildcardCls

type family SemaInfoClassify (node :: * -> * -> *) where ...

Equations

SemaInfoClassify UQualifiedName = SemaInfoNameCls

SemaInfoClassify ULiteral = SemaInfoLitCls

SemaInfoClassify UExpr = SemaInfoExprCls

SemaInfoClassify UImportDecl = SemaInfoImportCls

SemaInfoClassify UModule = SemaInfoModuleCls

SemaInfoClassify UFieldWildcard = SemaInfoWildcardCls

SemaInfoClassify a = SemaInfoDefaultCls

type family SemanticInfo' (domain :: *) (nodecls :: *)

type Domain d = (Typeable d, Data d, SemanticInfo' d SemaInfoDefaultCls ~ NoSemanticInfo, Data (SemanticInfo' d SemaInfoNameCls), Data (SemanticInfo' d SemaInfoLitCls), Data (SemanticInfo' d SemaInfoExprCls), Data (SemanticInfo' d SemaInfoImportCls), Data (SemanticInfo' d SemaInfoModuleCls), Data (SemanticInfo' d SemaInfoWildcardCls))

A semantic domain for the AST. The semantic domain maps semantic information for the different types of nodes in the AST. The kind of semantic domain for an AST depends on which stages of the compilation it passed. However after transforming the GHC representation to our AST, the domain stays the same. The domain is not applied to the AST elements that are generated while refactoring.

type DomainWith e d = (Data (SemanticInfo' d (SemaInfoClassify e)), Domain d)

class HasRange a where

Extracts or modifies the concrete range corresponding to a given source info. In case of lists and optional elements, it may not contain the elements inside.

Minimal complete definition

getRange, setRange

Methods

getRange :: a -> SrcSpan

setRange :: SrcSpan -> a -> a

class (Typeable stage, Data stage, Data (SpanInfo stage), Data (ListInfo stage), Data (OptionalInfo stage), HasRange (SpanInfo stage), HasRange (ListInfo stage), HasRange (OptionalInfo stage)) => SourceInfo stage

Class for source information stages

Associated Types

data SpanInfo stage :: *

UType of source info for normal AST elements

data ListInfo stage :: *

UType of source info for lists of AST elements

data OptionalInfo stage :: *

UType of source info for optional AST elements

shortShowSpan :: SrcSpan -> String

A short form of showing a range, without file name, for debugging purposes.

shortShowSpanWithFile :: SrcSpan -> String

shortShowLoc :: SrcLoc -> String

A short form of showing a range, without file name, for debugging purposes.

class SourceInfo stage => RangeInfo stage where

A class for marking a source information stage. All programs, regardless of correct Haskell programs or not, must go through these stages to be refactored.

Minimal complete definition

nodeSpan, listPos, optionalPos

Methods

nodeSpan :: Simple Lens (SpanInfo stage) SrcSpan

listPos :: Simple Lens (ListInfo stage) SrcLoc

optionalPos :: Simple Lens (OptionalInfo stage) SrcLoc

Annotations

data NodeInfo sema src

Semantic and source code related information for an AST node.

Constructors

NodeInfo

Fields

_semanticInfo :: sema

_sourceInfo :: src

sourceInfo :: forall sema src src'. Lens (NodeInfo sema src) (NodeInfo sema src') src src'

semanticInfo :: forall sema src sema'. Lens (NodeInfo sema src) (NodeInfo sema' src) sema sema'

data Ann elem dom stage

An element of the AST keeping extra information.

Constructors

Ann

Fields

_annotation :: NodeInfo (SemanticInfo dom elem) (SpanInfo stage)

The extra information for the AST part

_element :: elem dom stage

The original AST part

element :: forall elem dom stage. Lens (Ann elem dom stage) (Ann elem dom stage) (elem dom stage) (elem dom stage)

annotation :: forall elem dom stage. Lens (Ann elem dom stage) (Ann elem dom stage) (NodeInfo (SemanticInfo dom elem) (SpanInfo stage)) (NodeInfo (SemanticInfo dom elem) (SpanInfo stage))

data AnnListG elem dom stage

A list of AST elements

Constructors

AnnListG

Fields

_annListAnnot :: NodeInfo (SemanticInfo dom (AnnListG elem)) (ListInfo stage)

_annListElems :: [Ann elem dom stage]

annListElems :: forall elem dom stage. Lens (AnnListG elem dom stage) (AnnListG elem dom stage) [Ann elem dom stage] [Ann elem dom stage]

annListAnnot :: forall elem dom stage. Lens (AnnListG elem dom stage) (AnnListG elem dom stage) (NodeInfo (SemanticInfo dom (AnnListG elem)) (ListInfo stage)) (NodeInfo (SemanticInfo dom (AnnListG elem)) (ListInfo stage))

annList :: Traversal (AnnListG e d s) (AnnListG e d s) (Ann e d s) (Ann e d s)

data AnnMaybeG elem dom stage

An optional AST element

Constructors

AnnMaybeG

Fields

_annMaybeAnnot :: NodeInfo (SemanticInfo dom (AnnMaybeG elem)) (OptionalInfo stage)

_annMaybe :: Maybe (Ann elem dom stage)

annMaybeAnnot :: forall elem dom stage. Lens (AnnMaybeG elem dom stage) (AnnMaybeG elem dom stage) (NodeInfo (SemanticInfo dom (AnnMaybeG elem)) (OptionalInfo stage)) (NodeInfo (SemanticInfo dom (AnnMaybeG elem)) (OptionalInfo stage))

annMaybe :: forall elem dom stage. Lens (AnnMaybeG elem dom stage) (AnnMaybeG elem dom stage) (Maybe (Ann elem dom stage)) (Maybe (Ann elem dom stage))

class HasSourceInfo e where

Minimal complete definition

srcInfo

Associated Types

type SourceInfoType e :: *

Methods

srcInfo :: Simple Lens e (SourceInfoType e)

annJust :: Partial (AnnMaybeG e d s) (AnnMaybeG e d s) (Ann e d s) (Ann e d s)

annNil :: NodeInfo (SemanticInfo d (AnnListG e)) (ListInfo s) -> AnnListG e d s

An empty list of AST elements

isAnnNothing :: AnnMaybeG e d s -> Bool

isAnnJust :: AnnMaybeG e d s -> Bool

annLength :: AnnListG e d s -> Int

annNothing :: NodeInfo (SemanticInfo d (AnnMaybeG e)) (OptionalInfo s) -> AnnMaybeG e d s

A non-existing AST part

Info types

class ApplySemaChange cls where

A class for changing semantic information throught the AST.

Minimal complete definition

appSemaChange

Methods

appSemaChange :: SemaTrf f dom1 dom2 -> SemanticInfo' dom1 cls -> f (SemanticInfo' dom2 cls)

class ApplySemaChange (SemaInfoClassify a) => SemanticTraversal a where

A class for traversing semantic information in an AST

Minimal complete definition

semaTraverse

Methods

semaTraverse :: Monad f => SemaTrf f dom1 dom2 -> a dom1 st -> f (a dom2 st)

data SemaTrf f dom1 dom2

A transformation on the possible semantic informations for a given domain

Constructors

SemaTrf

Fields

trfSemaNameCls :: SemanticInfo' dom1 SemaInfoNameCls -> f (SemanticInfo' dom2 SemaInfoNameCls)

trfSemaExprCls :: SemanticInfo' dom1 SemaInfoExprCls -> f (SemanticInfo' dom2 SemaInfoExprCls)

trfSemaLitCls :: SemanticInfo' dom1 SemaInfoLitCls -> f (SemanticInfo' dom2 SemaInfoLitCls)

trfSemaImportCls :: SemanticInfo' dom1 SemaInfoImportCls -> f (SemanticInfo' dom2 SemaInfoImportCls)

trfSemaModuleCls :: SemanticInfo' dom1 SemaInfoModuleCls -> f (SemanticInfo' dom2 SemaInfoModuleCls)

trfSemaWildcardCls :: SemanticInfo' dom1 SemaInfoWildcardCls -> f (SemanticInfo' dom2 SemaInfoWildcardCls)

trfSemaDefault :: SemanticInfo' dom1 SemaInfoDefaultCls -> f (SemanticInfo' dom2 SemaInfoDefaultCls)

class SourceInfoTraversal a where

A class for traversing source information in an AST

Minimal complete definition

sourceInfoTraverseUp, sourceInfoTraverseDown, sourceInfoTraverse

Methods

sourceInfoTraverseUp :: Monad f => SourceInfoTrf f st1 st2 -> f () -> f () -> a dom st1 -> f (a dom st2)

sourceInfoTraverseDown :: Monad f => SourceInfoTrf f st1 st2 -> f () -> f () -> a dom st1 -> f (a dom st2)

sourceInfoTraverse :: Monad f => SourceInfoTrf f st1 st2 -> a dom st1 -> f (a dom st2)

data SourceInfoTrf f st1 st2

A transformation on the possible source informations

Constructors

SourceInfoTrf

Fields

trfSpanInfo :: SpanInfo st1 -> f (SpanInfo st2)

trfListInfo :: ListInfo st1 -> f (ListInfo st2)

trfOptionalInfo :: OptionalInfo st1 -> f (OptionalInfo st2)