Language.Haskell.Tools.AST.SemaInfoClasses

Documentation

semanticsLitType :: Ann ULiteral IdDom st -> Type

Information about names

type HasNameInfo dom = (Domain dom, HasNameInfo' (SemanticInfo dom UQualifiedName))

Domains that have semantic information for names

class HasNameInfo' si where

Infos that may have a name that can be extracted

Minimal complete definition

semanticsName

Methods

semanticsName :: si -> Maybe Name

Information about typed names

type HasIdInfo dom = (Domain dom, HasIdInfo' (SemanticInfo dom UQualifiedName))

class HasNameInfo' si => HasIdInfo' si where

Infos that may have a typed name that can be extracted

Minimal complete definition

semanticsId

Methods

semanticsId :: si -> Id

Fixity information

type HasFixityInfo dom = (Domain dom, HasFixityInfo' (SemanticInfo dom UQualifiedName))

class HasFixityInfo' si where

Infos that may have a fixity information

Minimal complete definition

semanticsFixity

Methods

semanticsFixity :: si -> Maybe Fixity

Scope information

type HasScopeInfo dom = (Domain dom, HasScopeInfo' (SemanticInfo dom UQualifiedName), HasScopeInfo' (SemanticInfo dom UExpr))

class HasScopeInfo' si where

Infos that contain the names that are available in theirs scope

Minimal complete definition

semanticsScope

Methods

semanticsScope :: si -> Scope

Information about names being defined

type HasDefiningInfo dom = (Domain dom, HasDefiningInfo' (SemanticInfo dom UQualifiedName))

class HasDefiningInfo' si where

Infos that store if they were used to define a name

Minimal complete definition

semanticsDefining

Methods

semanticsDefining :: si -> Bool

Information about source info in sema

class HasSourceInfoInSema' si where

Minimal complete definition

semanticsSourceInfo

Methods

semanticsSourceInfo :: si -> Maybe SrcSpan

Information about modules

type HasModuleInfo dom = (Domain dom, HasModuleInfo' (SemanticInfo dom UModule))

class HasModuleInfo' si where

Minimal complete definition

semanticsModule, semanticsDynFlags, isBootModule, semanticsImplicitImports, semanticsPrelInsts, semanticsPrelFamInsts

Methods

semanticsModule :: si -> Module

semanticsDynFlags :: si -> DynFlags

isBootModule :: si -> Bool

semanticsImplicitImports :: si -> [Name]

semanticsPrelInsts :: si -> [ClsInst]

semanticsPrelFamInsts :: si -> [FamInst]

Information about imports

type HasImportInfo dom = (Domain dom, HasImportInfo' (SemanticInfo dom UImportDecl))

class HasImportInfo' si where

Minimal complete definition

semanticsImportedModule, semanticsAvailable, semanticsImported, semanticsInsts, semanticsFamInsts

Methods

semanticsImportedModule :: si -> Module

semanticsAvailable :: si -> [Name]

semanticsImported :: si -> [Name]

semanticsInsts :: si -> [ClsInst]

semanticsFamInsts :: si -> [FamInst]

Information about implicitly bounded fields

type HasImplicitFieldsInfo dom = (Domain dom, HasImplicitFieldsInfo' (SemanticInfo dom UFieldWildcard))

class HasImplicitFieldsInfo' si where

Minimal complete definition

semanticsImplicitFlds

Methods

semanticsImplicitFlds :: si -> [(Name, Name)]

AST elements with no information

type HasNoSemanticInfo dom si = SemanticInfo dom si ~ NoSemanticInfo

data UsageSpec

Constructors

UsageSpec

Fields

usageQualified :: Bool

usageQualifier :: String

usageAs :: String